+++ src/cli/CHANGELOG.md
@@ -9,6 +9,10 @@ Section ordering: Added, Changed, Fixed, Removed, Security, Deprecated.
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `Header` layout components (and summary `componentType` refs) to `Heading` when running `studioctl app upgrade v9`.
+
 ## [0.1.0-preview.19] - 2026-08-06
 
 ### Added